:truck: api: move certification candidate for supervising from lib to src
