+++ api/tests/certification/supervising/integration/infrastructure/repositories/certification-candidate-for-supervising-repository_test.js
@@ -1,7 +1,7 @@
-import { catchErr, databaseBuilder, expect, knex, domainBuilder } from '../../../test-helper.js';
-import * as certificationCandidateForSupervisingRepository from '../../../../lib/infrastructure/repositories/certification-candidate-for-supervising-repository.js';
-import { CertificationCandidateNotFoundError } from '../../../../lib/domain/errors.js';
-import { Assessment } from '../../../../src/shared/domain/models/Assessment.js';
+import { catchErr, databaseBuilder, expect, knex, domainBuilder } from '../../../../../test-helper.js';
+import * as certificationCandidateForSupervisingRepository from '../../../../../../src/certification/supervision/infrastructure/repositories/certification-candidate-for-supervising-repository.js';
+import { CertificationCandidateNotFoundError } from '../../../../../../lib/domain/errors.js';
+import { Assessment } from '../../../../../../src/shared/domain/models/Assessment.js';
 
 describe('Integration | Repository | certification candidate for supervising', function () {
   describe('#get', function () {
