+++ api/tests/unit/domain/usecases/authorize-certification-candidate-to-start_test.js
@@ -1,6 +1,6 @@
 import { expect, sinon } from '../../../test-helper.js';
 import { authorizeCertificationCandidateToStart } from '../../../../lib/domain/usecases/authorize-certification-candidate-to-start.js';
-import { CertificationCandidateForSupervising } from '../../../../lib/domain/models/CertificationCandidateForSupervising.js';
+import { CertificationCandidateForSupervising } from '../../../../src/certification/supervision/domain/models/CertificationCandidateForSupervising.js';
 
 describe('Unit | Domain | Use Cases | authorize-certification-candidate-to-start', function () {
   let certificationCandidateForSupervisingRepository;
